pyproject.toml
[tool.poetry]
name = "agent_tools"
version = "0.1.0"
description = "General agent tool"
authors = ["happyPydog <lanc840930.mg09@nycu.edu.tw>"]
packages = [{ include = "agent_tools", from = "src" }]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13" 
pydantic = "^2.9.2"
openai = "^1.52.0"
magentic = "^0.32.0"
logfire = "^1.2.0"
tiktoken = "^0.8.0"
langfuse = "^2.52.1"
python-dotenv = "^1.0.1"
arxiv = "^2.1.3"
pymupdf = "^1.24.11"
more-itertools = "^10.5.0"
plotly = "^5.24.1"
scikit-learn = "^1.5.2"
langchain-core = "^0.3.12"
langchain-openai = "^0.2.3"
langchain-text-splitters = "^0.3.0"
langgraph = "^0.2.39"
unstructured = {extras = ["all-docs"], version = "^0.16.3"}
nbformat = "^5.10.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
isort = "^5.13.2"
mypy = "^1.12.0"
black = "^24.10.0"
flake8 = "^7.1.1"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
quiet = true


[tool.mypy]
ignore_missing_imports = true
warn_return_any = true
check_untyped_defs = true


[tool.flake8]
max-line-length = 120
indent-size = 4
ignore = ["E501", "E704"]

[tool.isort]
profile = "black"
